Tmr Ints disable/enable functionality
Tmr Ints disable/enable functionality:

-> Timer 1 channel 4 interrupts are enabled/disabled based on the Ignition Switch ON/OFF status respectively this is to avoid counting pulses when the Ignition switch is in OFF state
-> Odo Speedo pin is configured to Pull Down before it was is in floating state (No Pull state)
*****************************************************************************************************************
Modified prints
>for output visualization for Speed, Tacho, and Odo
*****************************************************************************************************************
Integrated the Tacho,Speedo and Odometer

*****************************************************************************************************************
BaselLine V0.2

Dhana changes Merged
-->DigitalDebounce
-->StateManager
-->IgnitionManager
--> UART Printf Configuration

*****************************************************************************************************************BaseLine_V0.1

MCAL Configurations
>Timers (Tim1 CH1& Tim4 CH4)
>Window Watchdog
>LTDC
>ADC(PA0 - Channel 1)
>RTC
>GPIO (MODE[PE3], RESET[PG5], IGN[PF8], UserButton -Switch_Fuel[PC13])

+++ STM32CubeIDE/Application/User/Core/src/App/Tachometer_App/Tachometer_App_cfg.h
@@ -56,7 +56,7 @@ typedef union {
         uint8_t fuel_bar_indicator : 1;
         uint8_t over_speed_indicator : 1;
     } indicators;
-    uint8_t all_indicators;
+    uint16_t all_indicators;
  }IndicationStatus_t;
  extern IndicationStatus_t IndicationStatus;
